ocr_fastapi_project/requirements.txt
fastapi
uvicorn
easyocr
torch
Pillow
vietocr
python-docx
opencv-python
